Update to Build.bat
Revised bat to simplify the process in getting files/folders in correct structure. This is to make it easier to simply copy to folder contained within "Build" to the required community location as a one time "copy and paste" action or further scripts may then use that to place the file as desired

May further update with default community locations if present however I do not use these myself so hasn't been implemented.